requirements.txt
flask 
faiss-cpu 
sentence-transformers 
numpy 
requests 
python-dotenv